refactor: import문 추상화한 형태로 사용
