+++ src/components/common/Icon/ReadumMarkLoadingIcon.tsx
@@ -1,5 +1,5 @@
 import { ReadumMarkLg } from '@/assets';
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib';
 
 type Props = {
   className?: string;
